% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhyloprofile.R
\name{runPhyloProfileCorona}
\alias{runPhyloProfileCorona}
\title{Run PhyloProfile app}
\usage{
runPhyloProfileCorona()
}
\value{
A shiny application - GUI version of PhyloProfile
}
\description{
Run PhyloProfile app
}
